user mailer for authentication token
perform sign_in/sign_out

improve/extend controller tests with authorization

show sign out when signed in

use authentication_token in email

finalizing branch

+++ app/controllers/tokens_controller.rb
@@ -1,23 +1,27 @@
 class TokensController < ApplicationController
-  def new
-    poll = find_poll
+  before_action :set_poll
 
-    @form = TokenForm.new(poll)
+  def new
+    authorize @poll, :manage?
+
+    @form = TokenForm.new(@poll)
   end
 
   def create
-    poll = find_poll
+    authorize @poll, :manage?
 
-    @form = TokenForm.new(poll, poll_token_params)
+    @form = TokenForm.new(@poll, poll_token_params)
 
     if @form.save
-      redirect_to poll
+      redirect_to @poll
     else
       render :new
     end
   end
 
   def destroy
+    authorize @poll, :manage?
+
     token = find_token
     token.destroy
     redirect_to token.poll
@@ -29,11 +33,11 @@ class TokensController < ApplicationController
     params.require(:poll_token).permit(:amount)
   end
 
-  def find_poll
-    Poll.find_by!(custom_id: params[:poll_custom_id])
+  def set_poll
+    @poll = Poll.find_by!(custom_id: params[:poll_custom_id])
   end
 
   def find_token
-    find_poll.tokens.find_by!(value: params[:value])
+    @poll.tokens.find_by!(value: params[:value])
   end
 end
